explanatory_gender_classification/requirements.txt
# Core PyTorch and Computer Vision
torch>=1.9.0
torchvision>=0.10.0
efficientnet-pytorch>=0.7.1

# Data Science and Machine Learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Image Processing
pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Explainability
lime>=0.2.0.1

# COCO Dataset (if using)
pycocotools>=2.0.4

# Progress Bars and Utilities
tqdm>=4.62.0

# Model Summary
torchsummary>=1.5.1

# Jupyter (for demo notebooks)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Additional utilities
requests>=2.26.0
Pillow>=8.3.0